RN-394: Converted @tupaia/data-api tests to typescript
