+++ package.json
@@ -101,6 +101,7 @@
     "sinon-chai": "^3.5.0",
     "start-server-and-test": "^1.11.0",
     "supertest": "^3.1.0",
+    "ts-node": "^10.7.0",
     "typescript": "^4.2.0",
     "winston": "^3.3.3",
     "yargs": "^16.0.3"
