+++ packages/data-api/.babelrc.js
@@ -1,0 +1,4 @@
+//TODO: Investigate how to do this via jest flag
+module.exports = {
+  extends: '../../.babelrc-ts.js',
+};
