+++ packages/data-api/jest.config.ts
@@ -1,7 +1,12 @@
-const baseConfig = require('../../jest.config-js.json');
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2022 Beyond Essential Systems Pty Ltd
+ */
+
+import baseConfig from '../../jest.config-ts.json';
 
 module.exports = async () => ({
   ...baseConfig,
   rootDir: '.',
-  setupFilesAfterEnv: ['../../jest.setup.js', './jest.setup.js'],
+  setupFilesAfterEnv: ['../../jest.setup.js', './jest.setup.ts'],
 });
