--- packages/data-api/src/__tests__/aggregations/index.ts
@@ -1,6 +1,0 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
- */
-
-export { testFetchAnalyticsAggregations } from './testFetchAnalyticsAggregations';
